Revert "[licensing] add license fetcher cache (#170006)"
This reverts commit 21c0b0b4205ba2d442656776b23b18e2dc062940.
