+++ x-pack/plugins/licensing/common/license_update.ts
@@ -17,7 +17,6 @@ import {
   takeUntil,
   finalize,
   startWith,
-  throttleTime,
 } from 'rxjs/operators';
 import { hasLicenseInfoChanged } from './has_license_info_changed';
 import type { ILicense } from './types';
@@ -30,15 +29,11 @@ export function createLicenseUpdate(
 ) {
   const manuallyRefresh$ = new Subject<void>();
 
-  const fetched$ = merge(
-    triggerRefresh$,
-    manuallyRefresh$.pipe(
-      throttleTime(1000, undefined, {
-        leading: true,
-        trailing: true,
-      })
-    )
-  ).pipe(takeUntil(stop$), exhaustMap(fetcher), share());
+  const fetched$ = merge(triggerRefresh$, manuallyRefresh$).pipe(
+    takeUntil(stop$),
+    exhaustMap(fetcher),
+    share()
+  );
 
   // provide a first, empty license, so that we can compare in the filter below
   const startWithArgs = initialValues ? [undefined, initialValues] : [undefined];
