+++ x-pack/plugins/licensing/tsconfig.json
@@ -15,8 +15,7 @@
     "@kbn/i18n",
     "@kbn/analytics-client",
     "@kbn/subscription-tracking",
-    "@kbn/core-analytics-browser",
-    "@kbn/logging-mocks"
+    "@kbn/core-analytics-browser"
   ],
   "exclude": ["target/**/*"]
 }
